requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
mysqlclient==2.2.4
pydantic==2.6.2
pydantic_core==2.16.3
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
